requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
jira==3.10.5
PyGithub==2.4.0
groq==1.2.0
